Moved patching methids into seperate utils script called counterfactual_patching.py in addition to general reorganising
