+++ comma_dataset.py
@@ -26,7 +26,7 @@ from circuitsvis.topk_samples import topk_samples
 from IPython.display import HTML, display
 import plotly.express as px
 import plotly.graph_objects as go
-from summarization_utils.circuit_analysis import get_logit_diff
+from summarization_utils.patching_metrics import get_logit_diff
 from summarization_utils.tokenwise_ablation import (
     compute_ablation_modified_loss,
     load_directions,
