+++ summarization_utils/patching_hooks.py
@@ -1,0 +1,48 @@
+from jaxtyping import Float
+from torch import Tensor
+
+
+def patch_pos_head_vector(
+    orig_head_vector: Float[Tensor, "batch pos head_index d_head"],
+    hook,
+    pos,
+    head_index,
+    patch_cache,
+):
+    """Patches a head vector at a given position and head index.
+
+    Args:
+        orig_head_vector (TT["batch", "pos", "head_index", "d_head"]): Original head activation vector.
+        hook (Hook): Hook to patch.
+        pos (int): Position to patch.
+        head_index (int): Head index to patch.
+        patch_cache (Dict[str, torch.Tensor]): Patch cache.
+
+    Returns:
+        TT["batch", "pos", "head_index", "d_head"]: Patched head vector.
+    """
+    orig_head_vector[:, pos, head_index, :] = patch_cache[hook.name][
+        :, pos, head_index, :
+    ]
+    return orig_head_vector
+
+
+def patch_head_vector(
+    orig_head_vector: Float[Tensor, "batch pos head_index d_head"],
+    hook,
+    head_index,
+    patch_cache,
+):
+    """Patches a head vector at a given head index.
+
+    Args:
+        orig_head_vector (TT["batch", "pos", "head_index", "d_head"]): Original head activation vector.
+        hook (Hook): Hook to patch.
+        head_index (int): Head index to patch.
+        patch_cache (Dict[str, torch.Tensor]): Patch cache.
+
+    Returns:
+        TT["batch", "pos", "head_index", "d_head"]: Patched head vector.
+    """
+    orig_head_vector[:, :, head_index, :] = patch_cache[hook.name][:, :, head_index, :]
+    return orig_head_vector
