+++ summarization_utils/patching_metrics.py
@@ -50,60 +50,6 @@ def get_final_non_pad_token(
     return final_token_logits
 
 
-# =============== VISUALIZATION UTILS ===============
-def visualize_tensor(tensor, labels, zmin=-1.0, zmax=1.0):
-    """Visualizes a 3D tensor as a series of heatmaps.
-
-    Args:
-        tensor (torch.Tensor): Tensor to visualize.
-        labels (List[str]): List of labels for each slice in the tensor.
-        zmin (float, optional): Minimum value for the color scale. Defaults to -1.0.
-        zmax (float, optional): Maximum value for the color scale. Defaults to 1.0.
-
-    Raises:
-        AssertionError: If the number of labels does not match the number of slices in the tensor.
-    """
-    assert (
-        len(labels) == tensor.shape[-1]
-    ), "The number of labels should match the number of slices in the tensor."
-
-    def plot_slice(selected_slice):
-        """Plots a single slice of the tensor."""
-        fig = go.FigureWidget(
-            data=go.Heatmap(
-                z=tensor[:, :, selected_slice].numpy(),
-                zmin=zmin,
-                zmax=zmax,
-                colorscale="RdBu",
-            ),
-            layout=go.Layout(
-                title=f"Slice: {selected_slice} - Step: {labels[selected_slice]}",
-                yaxis=dict(autorange="reversed"),
-            ),
-        )
-        return fig
-
-    def on_slider_change(change):
-        """Updates the plot when the slider is moved."""
-        selected_slice = change["new"]
-        fig = plot_slice(selected_slice)
-        output.clear_output(wait=True)
-        with output:
-            display(fig)
-
-    slider = widgets.IntSlider(
-        min=0, max=tensor.shape[2] - 1, step=1, value=0, description="Slice:"
-    )
-    slider.observe(on_slider_change, names="value")
-    display(slider)
-
-    output = widgets.Output()
-    display(output)
-
-    with output:
-        display(plot_slice(0))
-
-
 # =============== METRIC UTILS ===============
 def get_final_token_logits(
     logits: Float[Tensor, "batch *pos vocab"],
@@ -465,155 +411,6 @@ def cache_to_logit_diff(
         cache=cache,
         pos=pos,
     )
-
-
-# =============== PATCHING & KNOCKOUT UTILS ===============
-def patch_pos_head_vector(
-    orig_head_vector: Float[Tensor, "batch pos head_index d_head"],
-    hook,
-    pos,
-    head_index,
-    patch_cache,
-):
-    """Patches a head vector at a given position and head index.
-
-    Args:
-        orig_head_vector (TT["batch", "pos", "head_index", "d_head"]): Original head activation vector.
-        hook (Hook): Hook to patch.
-        pos (int): Position to patch.
-        head_index (int): Head index to patch.
-        patch_cache (Dict[str, torch.Tensor]): Patch cache.
-
-    Returns:
-        TT["batch", "pos", "head_index", "d_head"]: Patched head vector.
-    """
-    orig_head_vector[:, pos, head_index, :] = patch_cache[hook.name][
-        :, pos, head_index, :
-    ]
-    return orig_head_vector
-
-
-def patch_head_vector(
-    orig_head_vector: Float[Tensor, "batch pos head_index d_head"],
-    hook,
-    head_index,
-    patch_cache,
-):
-    """Patches a head vector at a given head index.
-
-    Args:
-        orig_head_vector (TT["batch", "pos", "head_index", "d_head"]): Original head activation vector.
-        hook (Hook): Hook to patch.
-        head_index (int): Head index to patch.
-        patch_cache (Dict[str, torch.Tensor]): Patch cache.
-
-    Returns:
-        TT["batch", "pos", "head_index", "d_head"]: Patched head vector.
-    """
-    orig_head_vector[:, :, head_index, :] = patch_cache[hook.name][:, :, head_index, :]
-    return orig_head_vector
-
-
-def ablate_top_head_hook(
-    z: Float[Tensor, "batch pos head_index d_head"], hook, head_idx=0
-):
-    """Hook to ablate the top head of a given layer.
-
-    Args:
-        z (TT["batch", "pos", "head_index", "d_head"]): Attention weights.
-        hook ([type]): Hook.
-        head_idx (int, optional): Head index to ablate. Defaults to 0.
-
-    Returns:
-        TT["batch", "pos", "head_index", "d_head"]: Attention weights.
-    """
-    z[:, -1, head_idx, :] = 0
-    return z
-
-
-def get_knockout_perf_drop(model, heads_to_ablate, clean_tokens, metric):
-    """Gets the performance drop for a given model and heads to ablate.
-
-    Args:
-        model (nn.Module): Model to knockout.
-        heads_to_ablate (List[Tuple[int, int]]): List of tuples of layer and head indices to knockout.
-        clean_tokens (Tensor): Clean tokens.
-        answer_token_indices (Tensor): Answer token indices.
-
-    Returns:
-        Tensor: Performance drop.
-    """
-    # Adds a hook into global model state
-    for layer, head in heads_to_ablate:
-        ablate_head_hook = partial(ablate_top_head_hook, head_idx=head)
-        model.blocks[layer].attn.hook_z.add_hook(ablate_head_hook)
-
-    ablated_logits, ablated_cache = model.run_with_cache(clean_tokens)
-    ablated_logit_diff = metric(ablated_logits)
-
-    return ablated_logit_diff
-
-
-@typechecked
-def project_to_subspace(
-    vectors: Float[Tensor, "... d_model"],
-    subspace: Float[Tensor, "d_model d_subspace"],
-) -> Float[Tensor, "... d_model"]:
-    assert vectors.shape[-1] == subspace.shape[0]
-    basis_projections = einops.einsum(
-        vectors, subspace, "... d_model, d_model d_subspace -> ... d_subspace"
-    )
-    summed_projections = einops.einsum(
-        basis_projections, subspace, "... d_subspace, d_model d_subspace -> ... d_model"
-    )
-    return summed_projections
-
-
-def create_cache_for_dir_patching(
-    clean_cache: ActivationCache,
-    corrupted_cache: ActivationCache,
-    sentiment_dir: Float[Tensor, "d_model *d_das"],
-    model: HookedTransformer,
-    device: torch.device = None,
-) -> ActivationCache:
-    """
-    We patch the sentiment direction from corrupt to clean
-    """
-    if device is None:
-        device = sentiment_dir.device
-    if sentiment_dir.ndim == 1:
-        sentiment_dir = sentiment_dir.unsqueeze(1)
-    assert sentiment_dir.ndim == 2
-    sentiment_dir: Float[Tensor, "d_model d_das"] = sentiment_dir / sentiment_dir.norm(
-        dim=0, keepdim=True
-    )
-    cache_dict = dict()
-    for act_name, clean_value in clean_cache.items():
-        is_result = act_name.endswith("result")
-        is_resid = (
-            act_name.endswith("resid_pre")
-            or act_name.endswith("resid_post")
-            or act_name.endswith("attn_out")
-            or act_name.endswith("mlp_out")
-        )
-        if is_resid or is_result:
-            clean_value = clean_value.to(device)
-            corrupt_value = corrupted_cache[act_name].to(device)
-
-            corrupt_proj: Float[Tensor, "... d_model"] = project_to_subspace(
-                corrupt_value,
-                sentiment_dir,
-            )
-            clean_proj: Float[Tensor, "... d_model"] = project_to_subspace(
-                clean_value,
-                sentiment_dir,
-            )
-            cache_dict[act_name] = corrupt_value + (clean_proj - corrupt_proj)
-        else:
-            # Only patch the residual stream
-            cache_dict[act_name] = corrupted_cache[act_name].to(device)
-
-    return ActivationCache(cache_dict, model)
 
 
 class PatchingMetric(Enum):
