+++ toy_datasets.py
@@ -30,7 +30,7 @@ from IPython.display import HTML, display
 import plotly.express as px
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
-from summarization_utils.circuit_analysis import get_logit_diff
+from summarization_utils.patching_metrics import get_logit_diff
 from summarization_utils.tokenwise_ablation import (
     compute_ablation_modified_loss,
     load_directions,
@@ -60,7 +60,6 @@ from summarization_utils.toy_datasets import (
     ToyDeductionTemplate,
     ToyBindingTemplate,
     ToyProfilesTemplate,
-    get_position_dict,
 )
 
 # %%
